requirements.txt
agno
yfinance
python-dotenv
openai
tweepy
tavily-python
openai
langchain
# Core dependencies
streamlit
pandas
numpy

# Database
supabase
psycopg2-binary



# Data processing and visualization
plotly
matplotlib

# Date/time handling
pytz

# Utilities
requests
pydantic